split components checkout
