+++ client/src/components/checkout/index.js
@@ -1,0 +1,4 @@
+export { default as CouponForm } from './CouponForm'
+export { default as Address } from './Address'
+export { default as ProductSummary } from './ProductSummary'
+export { default as CheckoutButtons } from './CheckoutButtons'
