+++ client/src/components/forms/index.js
@@ -5,4 +5,5 @@ export { default as FileUpload } from './FileUpload'
 export { default as ProductUpdateForm } from './ProductUpdateForm'
 export { default as Search } from './Search'
 export { default as Star } from './Star'
-export { default as Address } from './Address'
+export { default as Address } from '../checkout/Address'
+export { default as CouponForm } from '../checkout/CouponForm'
