+++ client/src/pages/Checkout.jsx
@@ -4,11 +4,10 @@ import {
   getUserCart,
   emptyUserCart,
   saveUserAddress,
-  applyCoupon,
   createCashOrder
 } from '../functions/user'
 import { toast } from 'react-toastify'
-import {Address} from '../components/forms/index'
+import {Address, CouponForm, ProductSummary, CheckoutButtons} from '../components/checkout/index'
 
 const Checkout = ({history}) => {
 
@@ -46,63 +45,6 @@ const Checkout = ({history}) => {
       }
     })
   }
-
-  const applyDiscountCoupon = () => {
-    //console.log('send coupon backend', coupon);
-    applyCoupon(user.token, coupon)
-      .then((res) => {
-        //console.log('APPLY COUPON RES', res.data);
-        //error
-        if (res.data.err) {
-          setDiscountError(res.data.err)
-          dispatch({
-            type: "COUPON_APPLIED",
-            payload: false
-          })
-        }
-
-        if (res.data) {
-          setTotalAfterDiscount(res.data)
-          dispatch({
-            type: "COUPON_APPLIED",
-            payload: true
-          })
-        }
-        
-      })
-  }
-
-  const showProductSummary = () => 
-    products.map((p, i) => (
-    <div key={i}>
-      <p>
-        {p.product.title} ({p.color}) x {p.count} = {(p.product.price * p.count).toLocaleString()}円
-      </p>
-    </div>
-  ))
-
-  const showApplyCoupon = () => (
-    <div className="col text-center">
-      <input 
-        value={coupon}
-        className="form-control"
-        type="text"
-        placeholder="クーポン名を入力"
-        onChange={(e) => {
-          setCoupon(e.target.value)
-          setDiscountError('')
-          setTotalAfterDiscount(0)
-        }}
-      />
-      <button
-        onClick={applyDiscountCoupon}
-        className="btn btn-primary btn-raised mt-2"
-        disabled={!coupon}
-      >
-        クーポンを適用
-      </button>
-    </div>
-  )
 
   const emptyCart = () => {
     // localStorage からcartを削除
@@ -169,54 +111,29 @@ const Checkout = ({history}) => {
             setSubAddress={setSubAddress}
           />
           <hr />
-          <h4 className="text-center">クーポンを使用</h4>
-          <br />
-          {discountError && <h5 className="text-danger text-center">クーポンが確認できません。</h5>}
-          {totalAfterdiscount > 0 && <h5 className="text-success text-center">クーポンが適用されました。</h5>}
-          <br/>
-          {showApplyCoupon()}
+          <CouponForm
+            coupon={coupon}
+            setCoupon={setCoupon}
+            discountError={discountError}
+            setDiscountError={setDiscountError}
+            totalAfterdiscount={totalAfterdiscount}
+            setTotalAfterDiscount={setTotalAfterDiscount}
+          />
           <br />
         </div>
         <div className="col-md-5">
-          <h4 className="text-danger text-center">ご注文内容の確認</h4>
-          <hr />
-          <p>現在{ products.length}つの商品が入っています</p>
-          <hr />
-          {showProductSummary()}
-          <hr />
-          <p>カートの合計：{ total.toLocaleString()}円</p>
-          {totalAfterdiscount > 0 && 
-            <p className="text-success ">
-              クーポン適用後の合計：{(totalAfterdiscount * 1).toLocaleString()}円
-            </p>
-          }
-          <div className="row">
-            <div className="col-md-6">
-              {COD ? (<button
-                className="btn btn-primary btn-raised"
-                disabled={!addressSaved || !products.length}
-                onClick={setCashOrder}
-              >
-                代金引換で購入する
-              </button>)
-                : (<button
-                className="btn btn-primary btn-raised"
-                disabled={!addressSaved || !products.length}
-                onClick={() => history.push('/payment')}
-              >
-                オンライン決済で購入する
-              </button>)}
-            </div>
-            <div className="col-md-6">
-              <button
-                onClick={emptyCart}
-                disabled={!products.length}
-                className="btn btn-danger btn-raised"
-              >
-                カートを空にする
-              </button>
-            </div>
-          </div>
+          <ProductSummary
+            products={products}
+            total={total}
+            totalAfterdiscount={totalAfterdiscount}
+          />
+          <CheckoutButtons
+            products={products}
+            emptyCart={emptyCart}
+            addressSaved={addressSaved}
+            setCashOrder={setCashOrder}
+            COD={COD}
+          />
         </div>
       </div>
     </div>
